pdfToWord/requirements.txt
Flask==2.3.3
python-docx==0.8.11
PyPDF2==3.0.1
reportlab==4.0.4
Werkzeug==2.3.7
gunicorn==21.2.0
pdf2docx==0.5.6 